Move get_graphql_content to tests/api/utils
