+++ tests/api/test_core.py
@@ -3,7 +3,7 @@ from unittest.mock import Mock
 
 import graphene
 from django.shortcuts import reverse
-from tests.utils import get_graphql_content
+from tests.api.utils import get_graphql_content
 
 from saleor.graphql.core.utils import clean_seo_fields, snake_to_camel_case
 from saleor.graphql.product import types as product_types
