+++ tests/api/test_shop.py
@@ -6,7 +6,7 @@ from django.shortcuts import reverse
 from django_countries import countries
 from saleor.core.permissions import MODELS_PERMISSIONS
 from saleor.site.models import Site
-from tests.utils import get_graphql_content
+from tests.api.utils import get_graphql_content
 
 from .utils import assert_no_permission
